Separate git operations
